+++ src/main/java/nextstep/subway/dto/line/LineRequest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.dto;
+package nextstep.subway.dto.line;
 
 import nextstep.subway.domain.line.Line;
