+++ src/main/java/nextstep/subway/dto/line/LineResponse.java
@@ -1,4 +1,4 @@
-package nextstep.subway.dto;
+package nextstep.subway.dto.line;
 
 import nextstep.subway.domain.line.Line;
 import nextstep.subway.domain.station.Station;
